refactor: update import paths in authService and teaService to utilize the new api module, improving code organization and maintainability
