+++ src/services/auth.js
@@ -1,4 +1,4 @@
-import api from '../utils/axios';
+import api from './api';
 
 export const authService = {
   login: (credentials) => api.post('/auth/login', credentials),
